removing incompatible unit tests
